capitulo9: Dielectrics funcionando

+++ hitable_list.h
@@ -15,7 +15,7 @@ class hitable_list: public hitable  {
 __device__ bool hitable_list::hit(const ray& r, float t_min, float t_max, hit_record& rec) const {
         hit_record temp_rec;
         bool hit_anything = false;
-        double closest_so_far = t_max;
+        float closest_so_far = t_max;
         for (int i = 0; i < list_size; i++) {
             if (list[i]->hit(r, t_min, closest_so_far, temp_rec)) {
                 hit_anything = true;
@@ -27,4 +27,3 @@ __device__ bool hitable_list::hit(const ray& r, float t_min, float t_max, hit_re
 }
 
 #endif
-
